Comments and background added to document
I’ve added comments about the background of each function. A link is
provided to the relevant page of R documentation which has blatantly
been copied here.

The routines all produce the same results to R with a test set of 50
data. This needs further testing to ensure accuracy but seems to be
reasonably close so far.