Unit tests for skewness and kurtosis - First commit
This is the first commit of the unit tests for the skewness and
kurtosis functions.

The first 2 test using a normal array for both functions. The next two
test the functions with an empty array (do they return ‘nil’
correctly?) and the third two analyse an array with a single element.

They probably need extending with other data sets.